refactor UserWorker in SaveUserRemote
